feat: add org open agent
