+++ schemas/org-open-agent.json
@@ -1,0 +1,22 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$ref": "#/definitions/OrgOpenOutput",
+  "definitions": {
+    "OrgOpenOutput": {
+      "type": "object",
+      "properties": {
+        "url": {
+          "type": "string"
+        },
+        "username": {
+          "type": "string"
+        },
+        "orgId": {
+          "type": "string"
+        }
+      },
+      "required": ["url", "username", "orgId"],
+      "additionalProperties": false
+    }
+  }
+}
