+++ src/shared/utils.ts
@@ -4,10 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { ChildProcess } from 'node:child_process';
 import { upperFirst } from '@salesforce/kit';
 import { StateAggregator } from '@salesforce/core';
-import open, { Options } from 'open';
 
 export const getAliasByUsername = async (username: string): Promise<string | undefined> => {
   const stateAggregator = await StateAggregator.getInstance();
@@ -16,8 +14,6 @@ export const getAliasByUsername = async (username: string): Promise<string | und
   return keys?.length ? keys[keys.length - 1] : undefined;
 };
 
-export const openUrl = async (url: string, options: Options): Promise<ChildProcess> => open(url, options);
-
 export const lowerToUpper = (object: Record<string, unknown>): Record<string, unknown> =>
   // the API has keys defined in capital camel case, while the definition schema has them as lower camel case
   // we need to convert lower camel case to upper before merging options so they will override properly
@@ -25,7 +21,6 @@ export const lowerToUpper = (object: Record<string, unknown>): Record<string, un
 
 export default {
   getAliasByUsername,
-  openUrl,
   lowerToUpper,
 };
